Initial commit
This is a CGI script used for mirroring a github hosted repo.
It's intended use is to be run as a github service hook The
usecase in mind is to be able to run local update hooks --- e.g.
for sending commit mails with diffs (something github
unfortunately doesn't support currently).